python/requirements.txt
# Yellowstone gRPC Client Dependencies

# gRPC and Protocol Buffers
grpcio~=1.74.0
grpcio-tools~=1.74.0

# Environment variables
python-dotenv~=1.1.1

# Reconnection logic with exponential backoff
backoff~=2.2.1